requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.10.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
django-environ==0.10.0
Flask==2.2.3
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.5.1
plotly==5.14.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.2.3
